+++ app/src/main/java/org/merakilearn/theme/Theme.kt
@@ -1,0 +1,42 @@
+package org.merakilearn.theme
+
+
+import java.util.*
+
+val currentDate: Calendar =  Calendar.getInstance()
+
+fun isChristmas(): Boolean {
+        val christmasStart = Calendar.getInstance().apply {
+            set(Calendar.MONTH, Calendar.DECEMBER)
+            set(Calendar.DAY_OF_MONTH, 22)
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+        }
+        val christmasEnd = Calendar.getInstance().apply {
+            set(Calendar.MONTH, Calendar.DECEMBER)
+            set(Calendar.DAY_OF_MONTH, 26)
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+        }
+        return currentDate.after(christmasStart) && currentDate.before(christmasEnd)
+}
+
+fun isNewYear(): Boolean {
+        val newYearStart = Calendar.getInstance().apply {
+            set(Calendar.MONTH, Calendar.DECEMBER)
+            set(Calendar.DAY_OF_MONTH, 30)
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+        }
+        val newYearEnd = Calendar.getInstance().apply {
+            set(Calendar.MONTH, Calendar.JANUARY)
+            set(Calendar.DAY_OF_MONTH,2 )
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+        }
+        return currentDate.after(newYearStart) && currentDate.before(newYearEnd)
+}
