+++ app/src/main/java/org/merakilearn/ui/SplashActivity.kt
@@ -9,13 +9,14 @@ import android.view.View
 import androidx.databinding.DataBindingUtil
 import org.merakilearn.R
 import org.merakilearn.databinding.ActivitySplashBinding
+import org.merakilearn.theme.isChristmas
+import org.merakilearn.theme.isNewYear
 import org.merakilearn.ui.onboarding.OnBoardingActivity
-import java.util.*
+
 
 
 class SplashActivity : AppCompatActivity() {
    private lateinit var binding : ActivitySplashBinding
-   private val currentDate =  Calendar.getInstance()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this,R.layout.activity_splash)
@@ -74,40 +75,4 @@ class SplashActivity : AppCompatActivity() {
         }
     }
 
-
-    private fun isChristmas(): Boolean {
-        val christmasStart = Calendar.getInstance().apply {
-            set(Calendar.MONTH, Calendar.DECEMBER)
-            set(Calendar.DAY_OF_MONTH, 22)
-            set(Calendar.HOUR_OF_DAY, 0)
-            set(Calendar.MINUTE, 0)
-            set(Calendar.SECOND, 0)
-        }
-        val christmasEnd = Calendar.getInstance().apply {
-            set(Calendar.MONTH, Calendar.DECEMBER)
-            set(Calendar.DAY_OF_MONTH, 26)
-            set(Calendar.HOUR_OF_DAY, 0)
-            set(Calendar.MINUTE, 0)
-            set(Calendar.SECOND, 0)
-        }
-        return currentDate.after(christmasStart) && currentDate.before(christmasEnd)
-    }
-
-    private fun isNewYear(): Boolean {
-        val newYearStart = Calendar.getInstance().apply {
-            set(Calendar.MONTH, Calendar.DECEMBER)
-            set(Calendar.DAY_OF_MONTH, 30)
-            set(Calendar.HOUR_OF_DAY, 0)
-            set(Calendar.MINUTE, 0)
-            set(Calendar.SECOND, 0)
-        }
-        val newYearEnd = Calendar.getInstance().apply {
-            set(Calendar.MONTH, Calendar.JANUARY)
-            set(Calendar.DAY_OF_MONTH,2 )
-            set(Calendar.HOUR_OF_DAY, 0)
-            set(Calendar.MINUTE, 0)
-            set(Calendar.SECOND, 0)
-        }
-        return currentDate.after(newYearStart) && currentDate.before(newYearEnd)
-    }
 }
